Finish sheet position

+++ src/cryngine/map/chunk.cr
@@ -4,8 +4,8 @@ module Cryngine
       class_getter width : UInt8 = 16.to_u8
       class_getter height : UInt8 = 16.to_u8
       getter data : Hash(UInt8, Hash(UInt8, Int16)) = Hash(UInt8, Hash(UInt8, Int16)).new
-      getter x : Int32
-      getter y : Int32
+      getter x : Int16
+      getter y : Int16
 
       def initialize(data, @x, @y)
         data.each_with_index do |sprite, i|
